@@ -14,8 +14,9 @@
     <name>spring-boot-exchange-rate</name>
     <description>spring-boot-exchange-rate</description>
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <spring-cloud.version>2021.0.1</spring-cloud.version>
+        <openapi-generator-version>7.10.0</openapi-generator-version>
     </properties>
     <dependencies>
         <dependency>
