@@ -4,7 +4,7 @@ package com.example.mibanco.exchangerate.models.entity;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 
 @Entity
